Groundwork for stuff + Create and delete nodes
Adding and removing nodes is now supported via two widgets on the left.
Click and drag the + down and right to make a new node, and drop nodes
on the x to remove them.
The internals are halfway through a major transition, this is the last
update until that transition is complete.